backend/auth/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.8.0
mysqlclient==2.2.4
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.3.0
python-decouple==3.8